pyproject.toml
[tool.poetry]
name = "fyers"
version = "0.1.0"
description = "Paper trading and backtesting system for Fyers"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fyers-apiv3 = "*" # This is a direct dependency
python-dotenv = "^1.0.0" # For loading .env files
pandas = "^2.3.2" # Core data manipulation library
pandas-ta = "^0.4.67b0" # Technical analysis library
pyyaml = "^6.0.1" # For handling YAML config files
schedule = "^1.2.2" # For scheduling tasks

# Streamlit and its ecosystem
streamlit = "^1.49.1"
streamlit-option-menu = "^0.4.0" # For the dashboard sidebar
streamlit-autorefresh = "^1.0.1" # For the live monitor
plotly = "^6.3.0" # For creating charts
psutil = "^7.0.0" # For checking process status

# Note: Other packages like requests, numpy, etc., will be installed
# as transitive dependencies of the packages listed above.

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[tool.pytest.ini_options]
pythonpath = [
  ".", "src"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"